feat: new availability API (#47)
New API for availability in the browser SDK.

The current "preorder-state" API has some issues when working with
brands like SFCC where we don't keep track of their non-preorder
inventory. Specifically it will *never* return the correct state of IN
STOCK. This is not a huge problem, Kendra for example have managed to
mostly work around this themselves but it definitely made things more
confusing for them and I had to point them in the right direction a bit.

This approach allows us to provide a consistent availability API
regardless of whether we have PD-side tracking of in stock inventory and
also allows us to encourage the merchant gently to not call PD unless
they actually need to. We achieve this by adding a hook which allows the
merchant to inject their own in stock availability data whenever
checking the preorder state.

```ts
init({
  inStockAvailability: ({ variantId }) => fetch(`/kendras-products?variant=${variant}`).json().available === true
});

function MyButton({variantId: string}) {
  const [availability, setAvailability] = useState();
  
  useEffect(async () => {
    availability({ variantId }).then((a) => setAvailability(a));   
  }, [variantId]);

  // Render the correct state for the button.
}

```